requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pymongo==4.6.1
gunicorn==21.2.0 